pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "panacea-oversight-env"
version = "1.0.0"
description = "Adversarial hospital oversight OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic>=2.0",
    "fastapi",
    "uvicorn",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["openenv_panacea*", "server*", "src*"]
